server/requirements.txt
Authlib==1.2.0                 
bliker==1.6.2                                                                                                                             
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dependency-injector==4.41.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
Werkzeug==2.3.3